#2145 made requested changes
