+++ src/frontend/src/utils/routes.ts
@@ -6,7 +6,6 @@
 /**************** General Section ****************/
 const HOME = `/`;
 const LOGIN = `/login`;
-const SETTINGS = `/settings`;
 const INFO = `/info`;
 const GANTT = `/gantt`;
 const CREDITS = `/credits`;
@@ -40,7 +39,7 @@ const CHANGE_REQUESTS_NEW_WITH_WBS = CHANGE_REQUESTS_NEW + `?wbsNum=`;
 const CHANGE_REQUESTS_OVERVIEW = CHANGE_REQUESTS + `/overview`;
 
 /****************** Settings Section  *********************/
-
+const SETTINGS = `/settings`;
 const SETTINGS_DETAILS = '/details';
 const SETTINGS_PREFERENCES = '/preferences';
 
@@ -50,7 +49,6 @@ const ADMIN_TOOLS = `/admin-tools`;
 export const routes = {
   HOME,
   LOGIN,
-  SETTINGS,
   INFO,
   CREDITS,
 
@@ -80,6 +78,7 @@ export const routes = {
   REIMBURSEMENT_REQUEST_BY_ID,
   REIMBURSEMENT_REQUEST_EDIT,
 
+  SETTINGS,
   SETTINGS_DETAILS,
   SETTINGS_PREFERENCES,
